Common Variant Genes

This function finds common variant genes between two cell data sets.

Usage

common_variant_genes(
  cds1,
  cds2,
  top_n = 2000,
  logmean_ul = 2,
  logmean_ll = -6,
  row_data_column = "gene_short_name",
  unique_data_column = "id",
  verbose = T,
  plot = F
)

Arguments

cds1

The first cell data set.

cds2

The second cell data set.

top_n

The number of top genes to consider. Default is 2000.

logmean_ul

The upper limit for mean expression.

logmean_ll

The lower limit for mean expression.

row_data_column

The column in the feature data corresponding to the gene symbol.

unique_data_column

The column in the feature data corresponding to the unique id.

verbose

Boolean indicating whether to display verbose output.

plot

Boolean indicating whether to plot the results.

Value

A vector of similarly variant genes.